+++ packages/storion/src/utils/storeTyple.ts
@@ -1,0 +1,11 @@
+import { ActionsBase, StateBase } from "../types";
+
+export function storeTuple<S extends StateBase, A extends ActionsBase>(
+  state: S,
+  actions: A
+) {
+  return Object.assign([state, actions] as const, {
+    state,
+    actions,
+  });
+}